Added new components, moved code from App to the respective components with minor updates
