+++ src/App.jsx
@@ -1,101 +1,20 @@
-import { useState } from 'react';
+import { useState, Fragment } from 'react';
 
-import { CORE_CONCEPTS } from './data.js';
 import Header from './components/Header.jsx';
 import CoreConcepts from './components/CoreConcepts.jsx';
-import TabButton from './components/TabButton.jsx';
-import { EXAMPLES } from './data.js';
+import Examples from './components/Examples.jsx';
 
 function App() {
-  const [selectedTopic, setSelectedTopic] = useState();
-
-  function handleSelect(selectedButton) {
-    //selectedButton => 'component', 'jsx', 'props', 'state'
-    setSelectedTopic(selectedButton);
-    //console.log(selectedTopic);
-  };
-
-  console.log('APP COMPONENT EXECUTING');
-
-  let tabContent = <p>Please select a topic.</p>;
-
-  if (selectedTopic) {
-    tabContent = (
-      <div id="tab-content">
-        <h3>{EXAMPLES[selectedTopic].title}</h3>
-        <p>{EXAMPLES[selectedTopic].description}</p>
-        <pre>
-          <code>
-            {EXAMPLES[selectedTopic].code}
-          </code>
-        </pre>
-      </div>
-    );
-  };
+  
 
   return (
-    <div>
+    <>
       <Header />
       <main>
-        <section id="core-concepts">
-          <h2>Core Concepts</h2>
-          <ul>
-            {/* Shortcut to display component data thats stored in an array, that might change in the future */}
-            {CORE_CONCEPTS.map((conceptItem) => (
-              <CoreConcepts key={conceptItem.title} {...conceptItem} />
-            ))}
-
-            {/* 
-            The code below does the same thing as the above code, except it hard codes the values in the data
-            so if the amount of the data changes it could *break* the program and not display correctly or not 
-            display all of the code thats actually in the array unless you hard code it in like below 
-            */}
-            {/* <CoreConcepts
-              title={CORE_CONCEPTS[0].title}
-              description={CORE_CONCEPTS[0].description}
-              image={CORE_CONCEPTS[0].image}
-            /> */}
-            {/* 
-            Shortcut for cutting down time on writting out the code for the component 
-            this method is still hard coded and not dynamic, meaning if the amount of data changes +/- 
-            it will only display what is coded below, or throw an error if the hard coded data  
-            no longer exists
-            */}
-            {/* <CoreConcepts {...CORE_CONCEPTS[1]} />
-            <CoreConcepts {...CORE_CONCEPTS[2]} />
-            <CoreConcepts {...CORE_CONCEPTS[3]} /> */}
-          </ul>
-
-        </section>
-        <section id="examples">
-          <h2>Examples</h2>
-          <menu>
-            <TabButton isSelected={selectedTopic === 'components'} onSelect={() => handleSelect('components')}>Components</TabButton>
-            <TabButton isSelected={selectedTopic === 'jsx'} onSelect={() => handleSelect('jsx')}>JSX</TabButton>
-            <TabButton isSelected={selectedTopic === 'props'} onSelect={() => handleSelect('props')}>Props</TabButton>
-            <TabButton isSelected={selectedTopic === 'state'} onSelect={() => handleSelect('state')}>State</TabButton>
-          </menu>
-
-
-          {/* 
-          One example of displaying content on the screen as well as if the state doesn't have a start value for this 
-          use case, below
-          */}
-          {/* {!selectedTopic ? <p>Please select a topic.</p> : (
-            <div id="tab-content">
-              <h3>{EXAMPLES[selectedTopic].title}</h3>
-              <p>{EXAMPLES[selectedTopic].description}</p>
-              <pre>
-                <code>
-                  {EXAMPLES[selectedTopic].code}
-                </code>
-              </pre>
-            </div>
-          )} */}
-          {tabContent}
-        </section>
+        <CoreConcepts />
+        <Examples />
       </main>
-    </div>
+    </>
   );
 }
