+++ src/components/Examples.jsx
@@ -1,0 +1,63 @@
+import { useState } from "react";
+
+import TabButton from "./TabButton.jsx";
+import { EXAMPLES } from "../data.js";
+
+
+export default function Examples() {
+    const [selectedTopic, setSelectedTopic] = useState();
+
+    function handleSelect(selectedButton) {
+      //selectedButton => 'component', 'jsx', 'props', 'state'
+      setSelectedTopic(selectedButton);
+      //console.log(selectedTopic);
+    };
+  
+    console.log('APP COMPONENT EXECUTING');
+  
+    let tabContent = <p>Please select a topic.</p>;
+  
+    if (selectedTopic) {
+      tabContent = (
+        <div id="tab-content">
+          <h3>{EXAMPLES[selectedTopic].title}</h3>
+          <p>{EXAMPLES[selectedTopic].description}</p>
+          <pre>
+            <code>
+              {EXAMPLES[selectedTopic].code}
+            </code>
+          </pre>
+        </div>
+      );
+    };
+
+    return (
+        <section id="examples">
+          <h2>Examples</h2>
+          <menu>
+            <TabButton isSelected={selectedTopic === 'components'} onSelect={() => handleSelect('components')}>Components</TabButton>
+            <TabButton isSelected={selectedTopic === 'jsx'} onSelect={() => handleSelect('jsx')}>JSX</TabButton>
+            <TabButton isSelected={selectedTopic === 'props'} onSelect={() => handleSelect('props')}>Props</TabButton>
+            <TabButton isSelected={selectedTopic === 'state'} onSelect={() => handleSelect('state')}>State</TabButton>
+          </menu>
+
+
+          {/* 
+          One example of displaying content on the screen as well as if the state doesn't have a start value for this 
+          use case, below
+          */}
+          {/* {!selectedTopic ? <p>Please select a topic.</p> : (
+            <div id="tab-content">
+              <h3>{EXAMPLES[selectedTopic].title}</h3>
+              <p>{EXAMPLES[selectedTopic].description}</p>
+              <pre>
+                <code>
+                  {EXAMPLES[selectedTopic].code}
+                </code>
+              </pre>
+            </div>
+          )} */}
+          {tabContent}
+        </section>
+    );
+}
